+++ tests/testthat/test-check_species.R
@@ -1,10 +1,10 @@
 test_that("Error is returned if species is NULL or of length 0", {
-  testthat::expect_error(check_species(mica, NULL))
-  testthat::expect_error(check_species(mica, character(0)))
+  expect_error(check_species(mica, NULL))
+  expect_error(check_species(mica, character(0)))
 })
 
 test_that("Error is returned if one or more species are invalid", {
-  testthat::expect_error(
+  expect_error(
     check_species(
       mica,
       c(
@@ -30,31 +30,31 @@ test_that("Error is returned if one or more species are invalid", {
 test_that("If input is a scientific name, the result is equal to input", {
   sc_name <- "Anas strepera"
   species <- check_species(mica, sc_name)
-  testthat::expect_equal(species, sc_name)
+  expect_equal(species, sc_name)
 })
 
 test_that("Multiput scientific names are allowed", {
   sc_names <- c("Anas strepera", "Ardea cinerea")
   species <- check_species(mica, sc_names)
-  testthat::expect_equal(species, sc_names)
+  expect_equal(species, sc_names)
 })
 
 test_that("Function works with vernacular names", {
   vn_names <- c("beech marten", "mallard")
   species <- suppressMessages(check_species(mica, vn_names))
-  testthat::expect_equal(species, c("Martes foina", "Anas platyrhynchos"))
+  expect_equal(species, c("Martes foina", "Anas platyrhynchos"))
 })
 
 test_that("Functions works well with vernacular names of different languages", {
   vn_names <- c("beech marten", "wilde eend")
   species <- suppressMessages(check_species(mica, vn_names))
-  testthat::expect_equal(species, c("Martes foina", "Anas platyrhynchos"))
+  expect_equal(species, c("Martes foina", "Anas platyrhynchos"))
 })
 
 test_that("Functions works with a mix of scientific and vernacular names", {
   mixed_names <- c("mallard", "steenmarter", "Castor fiber")
   species <- suppressMessages(check_species(mica, mixed_names))
-  testthat::expect_equal(
+  expect_equal(
     species,
     c(
       "Anas platyrhynchos",
@@ -66,11 +66,11 @@ test_that("Functions works with a mix of scientific and vernacular names", {
 
 test_that("Taxon IDs are not allowed", {
   taxon_id <- mica$taxonomic[[1]]$taxonID
-  testthat::expect_error(check_species(mica, taxon_id))
+  expect_error(check_species(mica, taxon_id))
 })
 
 test_that("Functions works case insensitively", {
   vn_name <- suppressMessages(check_species(mica, c("MallARD")))
   species <- check_species(mica, vn_name)
-  testthat::expect_equal(species, "Anas platyrhynchos")
+  expect_equal(species, "Anas platyrhynchos")
 })